requirements.txt
streamlit
chromadb
sentence-transformers
requests
python-dotenv
